+++ psm_utils/__init__.py
@@ -3,7 +3,6 @@
 __version__ = "0.9.0"
 __all__ = ["Peptidoform", "PSM", "PSMList"]
 
-from functools import lru_cache
 from warnings import filterwarnings
 
 # mzmlb is not used, so hdf5plugin is not needed
@@ -14,17 +13,6 @@ filterwarnings(
     module="psims.mzmlb",
 )
 
-from pyteomics.proforma import (  # noqa: E402
-    GenericResolver,
-    PSIModResolver,
-    UnimodResolver,
-)
-
 from psm_utils.peptidoform import Peptidoform  # noqa: E402
 from psm_utils.psm import PSM  # noqa: E402
 from psm_utils.psm_list import PSMList  # noqa: E402
-
-# Temporary patch until released in pyteomics (see levitsky/pyteomics#147)
-UnimodResolver.resolve = lru_cache(maxsize=None)(UnimodResolver.resolve)
-PSIModResolver.resolve = lru_cache(maxsize=None)(PSIModResolver.resolve)
-GenericResolver.resolve = lru_cache(maxsize=None)(GenericResolver.resolve)